First part of encoding refactor into transactions and ops.
