+++ zone-sdk/src/sequencer/actor.rs
@@ -573,7 +573,6 @@ mod tests {
         header::{ContentId, HeaderId},
         mantle::{
             MantleTx, Note, Op, SignedMantleTx, Transaction as _, Utxo,
-            encoding::Ops,
             ledger::{Inputs, Outputs},
             ops::{
                 OpProof,
@@ -585,6 +584,7 @@ mod tests {
                     withdraw::ChannelWithdrawOp,
                 },
             },
+            transactions::Ops,
         },
         proofs::leader_proof::Groth16LeaderProof,
     };
