+++ zone-sdk/src/sequencer/block_fetch.rs
@@ -587,7 +587,6 @@ mod tests {
     use lb_core::mantle::{
         MantleTx, Note,
         channel::{SlotTimeframe, SlotTimeout},
-        encoding::Ops,
         ledger::{Inputs, Outputs},
         ops::{
             OpId as _, OpProof,
@@ -598,6 +597,7 @@ mod tests {
                 withdraw::ChannelWithdrawOp,
             },
         },
+        transactions::Ops,
     };
     use lb_key_management_system_service::keys::{Ed25519Key, Ed25519Signature, ZkKey};
     use num_bigint::BigUint;
